feat: add PrintEventQr component for generating and printing event QR codes
