+++ src/app/(backoffice)/admin/settings/page.tsx
@@ -11,7 +11,7 @@ export default async function Page() {
   return (
     <div className='flex flex-col gap-4 py-4'>
       <UpdateFeatures />
-      <section className='mb-4'>
+      <section className='mb-4 space-y-4'>
         <div className='flex justify-between items-center px-4'>
           <h2 className='text-3xl font-bold text-accent'>Usuarios</h2>
           <CreateUserForm />
